zmon-worker/requirements.txt
dogpile.cache
functional
numpy
pysnmp
redis
requests
subprocess32
graphitesend
suds
dnspython
psycopg2
CherryPy
Jinja2
setproctitle
PyYAML
eventlog-writer==0.4.5
timeperiod2==0.1
######## MySQL-python (MySQLdb) is difficult to compile, PyMySQL is a pure python replacement 100% compatible
#MySQL-python
PyMySQL==0.6.1
pythonwhois==2.4.3
pymssql==2.1.1

#zalando-cmdb-client==0.9.1
#zalando-zompy==0.2.17
#zalando-ldapapi==0.2.7
#zalando-scmmanager==0.0.23
